최종 완성

+++ src/main/java/sangwon/wead/controller/CommentController.java
@@ -10,10 +10,11 @@ import org.springframework.web.bind.annotation.RequestParam;
 import sangwon.wead.resolover.annotation.Referer;
 import sangwon.wead.resolover.annotation.UserId;
 import sangwon.wead.service.CommentService;
+import sangwon.wead.service.DTO.CommentUploadDto;
 
 import static sangwon.wead.util.AlertPageRedirector.redirectAlertPage;
 
-/*
+
 @Controller
 @RequiredArgsConstructor
 public class CommentController {
@@ -25,27 +26,24 @@ public class CommentController {
                          @UserId String userId,
                          @Referer String referer,
                          Model model) {
+        if(content.isBlank()) return redirectAlertPage("댓글을 작성해주세요.", referer, model);
 
-        if(content.isBlank()) {
-            return redirectAlertPage("댓글을 작성해주세요.", referer, model);
-        }
-
-        CommentUploadForm commentUploadForm = CommentUploadForm.builder()
+        CommentUploadDto commentUploadDto = CommentUploadDto.builder()
                 .userId(userId)
                 .postId(postId)
                 .content(content)
                 .build();
-        commentService.uploadComment(commentUploadForm);
+        commentService.uploadComment(commentUploadDto);
         return "redirect:" + referer;
     }
 
     @PostMapping("/comments/{commentId}/delete")
     public String delete(@PathVariable("commentId") Long commentId,
+                         @UserId String userId,
                          @Referer String referer) {
-        // 댓글 삭제
+        commentService.checkCommentPermission(commentId, userId);
         commentService.deleteComment(commentId);
         return "redirect:" + referer;
     }
 
 }
-*/
