+++ src/main/java/sangwon/wead/controller/PostController.java
@@ -2,29 +2,25 @@ package sangwon.wead.controller;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
-import sangwon.wead.controller.param.PostFormParam;
+import sangwon.wead.controller.DTO.PostFormParam;
 import sangwon.wead.resolover.annotation.Referer;
 import sangwon.wead.resolover.annotation.UserId;
-import sangwon.wead.service.DTO.ListDto;
-import sangwon.wead.service.DTO.PostRowDto;
+import sangwon.wead.service.DTO.*;
 import sangwon.wead.service.ListService;
 import sangwon.wead.service.PostService;
 import sangwon.wead.service.CommentService;
+import sangwon.wead.service.UserService;
 import sangwon.wead.service.book.BookService;
 import sangwon.wead.service.book.search.BookSearchService;
 
 
 import static sangwon.wead.util.AlertPageRedirector.redirectAlertPage;
-
-
 
 
 @Controller
@@ -34,6 +30,7 @@ public class PostController {
     private final PostService postService;
     private final CommentService commentService;
     private final BookService bookService;
+    private final UserService userService;
     private final BookSearchService bookSearchService;
     private final ListService listService;
 
@@ -41,11 +38,10 @@ public class PostController {
 
     @GetMapping( "/posts")
     public String list(@RequestParam(value = "page", required = false, defaultValue = "1") int pageNumber,
-                           @RequestParam(value = "query", required = false, defaultValue = "") String query,
-                           @RequestParam(value = "search", required = false, defaultValue = "") String search,
-                           @Referer String referer,
-                           Model model) {
-
+                       @RequestParam(value = "query", required = false, defaultValue = "") String query,
+                       @RequestParam(value = "search", required = false, defaultValue = "") String search,
+                       @Referer String referer,
+                       Model model) {
         if(query.length() > 50) return redirectAlertPage("검색어는 최대 50자입니다.", referer, model);
 
         final String finalQuery = query;
@@ -59,7 +55,6 @@ public class PostController {
                 search = "";
             }
         }
-
         ListDto<PostRowDto> listDto = listService.builder(pageFactory)
                 .setUrl("/posts")
                 .setQuery("search", search)
@@ -71,65 +66,69 @@ public class PostController {
         return "page/post_list";
     }
 
-    /*
-
     @GetMapping("/posts/{postId}")
-    public String read(@UserId(required = false) String userId,
-                       @PathVariable("postId") Long postId,
+    public String read(@PathVariable("postId") Long postId,
+                       @RequestParam(value = "page", required = false, defaultValue = "1") int pageNumber,
                        Model model) {
         postService.increasePostViews(postId);
+        PostDto postDto = postService.getPost(postId);
+        UserDto userDto = userService.getUser(postDto.getUserId());
+        BookDto bookDto = bookService.getBook(postDto.getIsbn());
 
-        PostInfo postInfo = postService.getPostInfo(postId);
-        BookInfo bookInfo = bookService.getBookInfo(postInfo.getIsbn());
-        List<CommentInfo> commentInfoList = commentService.getCommentInfoList(postId);
+        ListDto<CommentRowDto> listDto = listService.builder(pageable -> commentService.getCommentByPostId(pageable, postId))
+                .setUrl("/posts/" + postId)
+                .build(PageRequest.of(pageNumber, 10, Sort.by("commentId")));
 
-        model.addAttribute("sectionModel", modelBuilder.buildPostReadModel(postInfo, bookInfo, commentInfoList, userId));
+        model.addAttribute("post", postDto);
+        model.addAttribute("user", userDto);
+        model.addAttribute("book", bookDto);
+        model.addAttribute("comment", listDto);
         return "page/post_read";
     }
 
     @GetMapping("/books/{isbn}/posts/upload")
     public String uploadForm(@PathVariable("isbn") String isbn, Model model) {
-        BookInfo bookInfo = bookSearchService.getBookInfo(isbn);
-        PostFormParam postFormParam = PostFormParam.builder().build();
-        model.addAttribute("sectionModel", modelBuilder.buildPostFormModel(bookInfo, postFormParam, false));
+        BookDto bookDto = bookSearchService.getBook(isbn);
+        model.addAttribute("book", bookDto);
+        model.addAttribute("type", "upload");
         return "page/post_form";
     }
+
 
     @PostMapping("/books/{isbn}/posts/upload")
     public String upload(@PathVariable("isbn") String isbn,
                          @Valid @ModelAttribute PostFormParam postFormParam,
                          BindingResult bindingResult,
                          @UserId String userId,
-                         Model model,
-                         RedirectAttributes redirectAttributes) {
+                         Model model) {
+        BookDto bookDto = bookSearchService.getBook(isbn);
 
         if(bindingResult.hasErrors()) {
-            BookInfo bookInfo = bookSearchService.getBookInfo(isbn);
-            model.addAttribute("sectionModel", modelBuilder.buildPostFormModel(bookInfo, postFormParam, true));
+            model.addAttribute("book", bookDto);
+            model.addAttribute("type", "upload");
+            model.addAttribute("title", postFormParam.getTitle());
+            model.addAttribute("content", postFormParam.getContent());
+            model.addAttribute("alert", true);
             return "page/post_form";
         }
 
-        if(!bookService.checkBookExistence(isbn)) {
-            BookInfo bookInfo = bookSearchService.getBookInfo(isbn);
-            bookService.saveBook(bookInfo);
-        }
-        Long postId = postService.uploadPost(postFormParam.toPostUploadForm(userId, isbn));
-
-        redirectAttributes.addAttribute("postId", postId);
-        return "redirect:/posts/{postId}";
+        bookService.saveBook(bookDto);
+        Long postId = postService.uploadPost(postFormParam.toPostUploadDto(userId, isbn));
+        return "redirect:/posts/" + postId;
     }
 
     @GetMapping("/posts/{postId}/update")
     public String updateForm(@PathVariable("postId") Long postId,
+                             @UserId String userId,
                              Model model) {
-        PostInfo postInfo = postService.getPostInfo(postId);
-        BookInfo bookInfo = bookService.getBookInfo(postInfo.getIsbn());
-        PostFormParam postFormParam = PostFormParam.builder()
-                .title(postInfo.getTitle())
-                .content(postInfo.getContent())
-                .build();
+        postService.checkPostPermission(postId, userId);
+        PostDto postDto = postService.getPost(postId);
+        BookDto bookDto = bookService.getBook(postDto.getIsbn());
 
-        model.addAttribute("sectionModel", modelBuilder.buildPostFormModel(bookInfo, postFormParam, false, postId));
+        model.addAttribute("book", bookDto);
+        model.addAttribute("type", "update");
+        model.addAttribute("title", postDto.getTitle());
+        model.addAttribute("content", postDto.getContent());
         return "page/post_form";
     }
 
@@ -137,29 +136,32 @@ public class PostController {
     public String update(@PathVariable("postId") Long postId,
                          @Valid @ModelAttribute PostFormParam postFormParam,
                          BindingResult bindingResult,
-                         Model model,
-                         RedirectAttributes redirectAttributes) {
+                         @UserId String userId,
+                         Model model) {
+        postService.checkPostPermission(postId, userId);
+        PostDto postDto = postService.getPost(postId);
+        BookDto bookDto = bookService.getBook(postDto.getIsbn());
 
         if(bindingResult.hasErrors()) {
-            PostInfo postInfo = postService.getPostInfo(postId);
-            BookInfo bookInfo = bookService.getBookInfo(postInfo.getIsbn());
-            model.addAttribute("sectionModel", modelBuilder.buildPostFormModel(bookInfo, postFormParam, true, postId));
+            model.addAttribute("book", bookDto);
+            model.addAttribute("type", "update");
+            model.addAttribute("title", postFormParam.getTitle());
+            model.addAttribute("content", postFormParam.getContent());
+            model.addAttribute("alert", true);
             return "page/post_form";
         }
 
-        postService.updatePost(postFormParam.toPostUpdateForm(postId));
-        redirectAttributes.addAttribute("postId", postId);
-        return "redirect:/posts/{postId}";
+        postService.updatePost(postFormParam.toPostUpdateDto(postId));
+        return "redirect:/posts/" + postId;
     }
 
     @PostMapping("/posts/{postId}/delete")
-    public String delete(@PathVariable("postId") Long postId) {
+    public String delete(@PathVariable("postId") Long postId,
+                         @UserId String userId) {
+        postService.checkPostPermission(postId, userId);
         postService.deletePost(postId);
         return "redirect:/posts";
     }
-
-     */
-
 }
